Add preference to get export filename from object
Add an addon preferences panel that contains a single preference to use
the active object's name as the initial filename instead of the blend
file's name.

The preference defaults to disabled, which maintains the existing
behaviour that matches default exporters - when first invoked the export
filepath is initialised to that of the blend file, with subsequent
exports remembering the previous filepath.

When the preference is enabled, the stem of the export filepath is set
to the active object's name each time the exporter is invoked - the rest
of the filepath behaves as before. Characters that are invalid in
Windows filenames are removed on all systems, with a placeholder being
used if the active object's name contains no valid characters.

Closes #2