Fix every material slot creating a new uv layer
The intention of the previous code was to create a new uv layer if the
mesh had no active uv layers, however the actual functionality was to
pro-actively create a new uv layer for every material slot. Blender's
limit of 8 uv layers was easily reached, resulting in an exception.

The new behaviour is to directly check if the mesh has no uv layers as
the first step, only creating a new uv layer in that case.

Fixes #4