Change main branch support level to 'COMMUNITY'
Starting with Blender 3.5, when a Blender version progresses into
development stage `bcon3` (Beta) the "Testing" tab is disabled, which
prevents enabling, disabling or adjusting addon preferences for addons
with support level `TESTING`. As a result, any user that tried to
install a copy of this addon from the `main` branch would be unable to
enable it (e.g. if they downloaded the repository as ZIP from GitHub).

The first commit on each release branch was already making this change,
so users downloading official releases are unaffected.