Fix split normals not being exported
Blender's own `Mesh.transform()` and `Mesh.flip_normals()` functions
destroy custom split normals, which is rather unhelpful. Converting to a
`bmesh` does not help. As there doesn't seem to be a way to use existing
Blender methods to perform the transformation and face winding reversal
on a whole mesh without destroying the custom split normals, the easiest
to implement solution I have chosen is to delay transformation to apply
to individual position and normal vectors at the extraction and
conversion phase, and winding reversal to apply to individual triangles,
even though this isn't as clean or efficient.

Finally loop normals are used in place of vertex normals, which should
have been the only change necessary.

Fixes #10