Fix diffuse colour for textured materials
The diffuse colour for materials that have an Image Texture node will
now always be opaque white, as was documented behaviour since initial
release but not actually implemented in the code. This prevents the
colour set when the Base Color input is disconnected from affecting the
export result even while hidden when an Image Texture node is connected.

If you have a use case for setting the diffuse colour of a textured
material to something other than opaque white please let me know about
it, preferably by leaving a comment on GitHub issue #12.

Fixes #12